Fix IA_ASM lint using correct x86_64 cross-compilation config
Previous runs used a local IA_ASM build without CMAKE_OSX_ARCHITECTURES=x86_64,
which is the flag CI uses. That flag enables portability-simd-intrinsics warnings
(all _mm_* calls flagged as non-portable), which were invisible locally.

Suppress portability-simd-intrinsics globally in .clang-tidy — the IA_ASM provider
is explicitly an x86 SIMD implementation, same rationale as reinterpret-cast.

Fix the remaining four non-simd warnings in IA_ASM files:
- rsa_oaep.hpp: hicpp-signed-bitwise (uint8_t XOR promotes to int) and
  readability-implicit-bool-conversion — rewrite constant-time mask using
  explicit unsigned arithmetic throughout
- rsa_pss.hpp: same hicpp-signed-bitwise pattern in constant-time compare
- sha256.hpp: misc-definitions-in-headers on [[gnu::target]] function (must
  live in header) — add NOLINT
- keccak.hpp: readability-function-size on fully-unrolled Keccak-f1600 — add NOLINT

Verified with build-lint-ia-x86 (CMAKE_OSX_ARCHITECTURES=x86_64), matching
the exact CI configuration. All six lint runs now produce zero warnings.
512 tests pass with no regressions.

+++ providers/ia_asm/rsa_pss.hpp
@@ -171,7 +171,7 @@ inline bool pss_verify(
     // Constant-time compare H' with H.
     uint8_t diff = 0U;
     for (std::size_t i = 0; i < oaep_hash_len; ++i) {
-        diff |= h[i] ^ h_prime[i];
+        diff |= static_cast<uint8_t>(static_cast<unsigned>(h[i]) ^ static_cast<unsigned>(h_prime[i]));
     }
     return diff == 0U;
 }
